requirements.txt
tensorflow-intel
keras
numpy
pandas
scikit-learn
joblib
matplotlib
fastapi
uvicorn
streamlit
